Try fixing mismatched foreign key issue on MySQL
Column `customer_id` on table `pay_charges` does not match column `id` on `pay_customers`, which has type `bigint`. To resolve this issue, change the type of the `customer_id` column on `pay_charges` to be :bigint. (For example `t.bigint :customer_id`).

Trying to updating migrations to 5.0 when the default changed to bigint