Allow syncing customers by overriding `pay_should_sync_customer?`
Before this change it was possible to trigger the
`enqeue_sync_email_job` method by changing the
`pay_should_sync_customer?` but the method would be a no-op unless the
email changed.

This commit makes it possible for someone to override _when_ the
customer would be synced by overriding the `pay_should_sync_customer?`
method.