Make a full request to sync first charge in a subscription
When a new subscription is created, we want to sync the first charge
before waiting for webhooks to process.

We aren't able to expand more than 4 levels deep with the Stripe API.
This changes it so we ask the API for the charge details instead of trying to expand everything in the subscribe request.